flask: Moved blueprints into top-level "views" folder
